requirements.txt
Django==1.11
psycopg2==2.7.6.1
django-mptt==0.9.0
django-merged-inlines==1.0.1
django-jet==1.0.8
django-nested-admin==3.1.2